+++ src/main/java/open/commons/spring/web/beans/authority/IRequestAccessAuthorityProvider.java
@@ -26,7 +26,7 @@
 
 package open.commons.spring.web.beans.authority;
 
-import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
 import org.springframework.stereotype.Controller;
@@ -71,6 +71,6 @@ public interface IRequestAccessAuthorityProvider extends IResourceAccessAuthorit
      * @since 2025. 5. 19.
      * @version 0.8.0
      */
-    Result<Boolean> isAllowed(@NotNull RequestMethod requestMethod, @NotEmpty String path);
+    Result<Boolean> isAllowed(@NotNull RequestMethod requestMethod, @NotBlank String path);
 
 }