+++ src/main/java/open/commons/spring/web/beans/resolver/IAuthorizedDataResolver.java
@@ -26,7 +26,7 @@
 
 package open.commons.spring.web.beans.resolver;
 
-import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotBlank;
 
 import org.springframework.beans.BeansException;
 import org.springframework.context.ApplicationContext;
@@ -69,7 +69,7 @@ public interface IAuthorizedDataResolver extends HandlerMethodArgumentResolver {
      * @since 2025. 9. 19.
      * @version 0.8.0
      */
-    default Object restoreValue(ApplicationContext context, String handleBean, @NotEmpty String handleType,
+    default Object restoreValue(ApplicationContext context, String handleBean, @NotBlank String handleType,
             Object rawValue) throws BeansException {
         if (rawValue == null) {
             return null;